+++ adapter-out/vancouver-opendata-client-adapter-out/build.gradle.kts
@@ -1,0 +1,6 @@
+dependencies {
+    implementation(project(":adapter-out:feign-adapter-out"))
+    implementation(project(":domain"))
+
+    implementation(project(":port-out:parking-worker-port-out"))
+}
