+++ adapter-out/vancouver-opendata-client-adapter-out/src/main/resources/vancouver-opendata-client-adapter-out.yml
@@ -1,0 +1,8 @@
+spring:
+    profiles:
+        include:
+            - feign-adapter-out
+
+client:
+    vancouver-opendata:
+        host: https://opendata.vancouver.ca
