+++ adapter-out/vancouver-opendata-client-adapter-out/src/test/resources/application.yml
@@ -1,0 +1,9 @@
+spring:
+    profiles:
+        include:
+            - feign-adapter-out
+
+client:
+    vancouver-opendata:
+        host: https://opendata.vancouver.ca
+
