--- adapter-out/vancouver-public-client-adapter-out/build.gradle.kts
@@ -1,4 +1,0 @@
-dependencies {
-    implementation(project(":adapter-out:feign-adapter-out"))
-    implementation(project(":domain"))
-}
