--- adapter-out/vancouver-public-client-adapter-out/src/main/resources/vancouver-public-client-adapter-out.yml
@@ -1,9 +1,0 @@
-# exclude config profile and redefine properties here
-spring:
-    profiles:
-        include:
-            - feign-adapter-out
-
-client:
-    vancouver-public:
-        host: https://opendata.vancouver.ca
