--- adapter-out/vancouver-public-client-adapter-out/src/test/resources/application.yml
@@ -1,8 +1,0 @@
-spring:
-    profiles:
-        include:
-            - feign-adapter-out
-
-client:
-    vancouver-public:
-        host: https://opendata.vancouver.ca
