+++ application/usecase-api/build.gradle.kts
@@ -1,5 +1,6 @@
 dependencies {
     implementation(project(":port-out:persistence-port-out"))
+    implementation(project(":port-out:parking-worker-port-out"))
     implementation(project(":port-out:mcp-port-out"))
 
     implementation(project(":port-in:usecase-api-port-in"))
