+++ port-out/parking-worker-port-out/build.gradle.kts
@@ -1,0 +1,3 @@
+dependencies {
+    implementation(project(":domain"))
+}
